+++ Tests/Unit/NetworkingDsc.Common.tests.ps1
@@ -106,37 +106,20 @@ try
             $adapterArray = @( $nomatchAdapter, $matchAdapter )
             $multipleMatchingAdapterArray = @( $matchAdapter, $matchAdapter )
 
-            Context 'No parameters are passed' {
-                Mock `
-                    -CommandName Get-NetAdapter `
-                    -MockWith { $adapterArray }
-
-                $errorRecord = Get-InvalidOperationRecord `
-                    -Message ($LocalizedData.NetAdapterParameterError)
-
-                It 'should throw exception' {
-                    { $script:result = Find-NetworkAdapter -Verbose } | Should Throw $errorRecord
-                }
-
-                It 'should call expected mocks' {
-                    Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 0
-                }
-            }
-
             Context 'Name is passed and one adapter matches' {
                 Mock `
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -Name $adapterName -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -149,11 +132,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -Name 'NOMATCH' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -163,15 +146,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -PhysicalMediaType $adapterPhysicalMediaType -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -184,11 +167,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -PhysicalMediaType 'NOMATCH' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -198,15 +181,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -Status $adapterStatus -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -219,11 +202,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -Status 'Disabled' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -233,15 +216,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -MacAddress $adapterMacAddress -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -254,11 +237,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -MacAddress '00-00-00-00-00-00' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -268,15 +251,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceDescription $adapterInterfaceDescription -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -289,11 +272,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceDescription 'NOMATCH' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -303,15 +286,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceIndex $adapterInterfaceIndex -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -324,11 +307,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceIndex 99 -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -338,15 +321,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceGuid $adapterInterfaceGuid -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -359,11 +342,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -InterfaceGuid 'NOMATCH' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -373,15 +356,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $adapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -DriverDescription $adapterDriverDescription -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -394,11 +377,46 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.NetAdapterNotFoundError)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -DriverDescription 'NOMATCH' -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
+                    Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
+                }
+            }
+
+            Context 'No parameters are passed and multiple Adapters adapters match but IgnoreMultipleMatchingAdapters is not set' {
+                Mock `
+                    -CommandName Get-NetAdapter `
+                    -MockWith { $adapterArray }
+
+                $errorRecord = Get-InvalidOperationRecord `
+                    -Message ($LocalizedData.MultipleMatchingNetAdapterFound -f 2)
+
+                It 'Should throw the correct exception' {
+                    { $script:result = Find-NetworkAdapter -Verbose } | Should Throw $errorRecord
+                }
+
+                It 'Should call expected mocks' {
+                    Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
+                }
+            }
+
+            Context 'No parameters are passed and multiple Adapters adapters match and IgnoreMultipleMatchingAdapters is set and interface number is 2' {
+                Mock `
+                    -CommandName Get-NetAdapter `
+                    -MockWith { $adapterArray }
+
+                It 'Should not throw exception' {
+                    { $script:result = Find-NetworkAdapter -IgnoreMultipleMatchingAdapters:$true -InterfaceNumber 2 -Verbose } | Should Not Throw
+                }
+
+                It 'Should return expected adapter' {
+                    $script:result.Name | Should Be $adapterName
+                }
+
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -411,11 +429,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.MultipleMatchingNetAdapterFound -f 2)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -PhysicalMediaType $adapterPhysicalMediaType -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -425,15 +443,15 @@ try
                     -CommandName Get-NetAdapter `
                     -MockWith { $multipleMatchingAdapterArray }
 
-                It 'should not throw exception' {
+                It 'Should not throw exception' {
                     { $script:result = Find-NetworkAdapter -PhysicalMediaType $adapterPhysicalMediaType -IgnoreMultipleMatchingAdapters:$true -Verbose } | Should Not Throw
                 }
 
-                It 'should return expected adapter' {
+                It 'Should return expected adapter' {
                     $script:result.Name | Should Be $adapterName
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }
@@ -446,11 +464,11 @@ try
                 $errorRecord = Get-InvalidOperationRecord `
                     -Message ($LocalizedData.InvalidNetAdapterNumberError -f 2,3)
 
-                It 'should throw exception' {
+                It 'Should throw the correct exception' {
                     { $script:result = Find-NetworkAdapter -PhysicalMediaType $adapterPhysicalMediaType -IgnoreMultipleMatchingAdapters:$true -InterfaceNumber 3 -Verbose } | Should Throw $errorRecord
                 }
 
-                It 'should call expected mocks' {
+                It 'Should call expected mocks' {
                     Assert-MockCalled -CommandName Get-NetAdapter -Exactly -Times 1
                 }
             }